Move `MaxpData`
