+++ src/lib.rs
@@ -86,6 +86,7 @@ mod remapper;
 mod write;
 
 use crate::interjector::Interjector;
+use crate::maxp::MaxpData;
 use crate::read::{Readable, Reader};
 pub use crate::remapper::GlyphRemapper;
 use crate::write::{Writeable, Writer};
@@ -339,7 +340,7 @@ struct Context<'a> {
     /// Normally, when subsetting CFF/TrueType fonts, we will simply read the corresponding
     /// data from the old font and rewrite it to the new font, for example when writing the
     /// `hmtx` table.
-    /// 
+    ///
     /// However, in case we are subsetting with variation coordinates, we instead rely on skrifa
     /// to apply the variation coordinates and interject that data during the subsetting process
     /// instead of using the data from the old font.
@@ -606,41 +607,3 @@ impl Display for Error {
 }
 
 impl std::error::Error for Error {}
-
-// When converting CFF2 to TTF, we need to write a version 1.0 MAXP table.
-pub(crate) struct MaxpData {
-    pub(crate) max_points: u16,
-    pub(crate) max_contours: u16,
-    pub(crate) max_composite_points: u16,
-    pub(crate) max_composite_contours: u16,
-    pub(crate) max_zones: u16,
-    pub(crate) max_twilight_points: u16,
-    pub(crate) max_storage: u16,
-    pub(crate) max_function_defs: u16,
-    pub(crate) max_instruction_defs: u16,
-    pub(crate) max_stack_elements: u16,
-    pub(crate) max_size_of_instructions: u16,
-    pub(crate) max_component_elements: u16,
-    pub(crate) max_component_depth: u16,
-}
-
-impl Default for MaxpData {
-    fn default() -> Self {
-        Self {
-            max_points: 0,
-            max_contours: 0,
-            max_composite_points: 0,
-            max_composite_contours: 0,
-            max_zones: 1,
-            max_twilight_points: 0,
-            max_storage: 0,
-            max_function_defs: 0,
-            max_instruction_defs: 0,
-            max_stack_elements: 0,
-            max_size_of_instructions: 0,
-            max_component_elements: 0,
-            // Could probably be 0 too since we only use simple glyphs when converting?
-            max_component_depth: 1,
-        }
-    }
-}
